pyproject.toml
[tool.poetry]
name = "ez_config_mgt"
description='Easy config management'
authors = ["Christophe Druet <christophe.druet@entsoe.eu>"]
version = "0.1.1"
license = "MIT"
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Development Status :: 2 - Pre-Alpha",
]
keywords = ["config", "management"]
packages = [{include = "ez_config_mgt"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
loguru = "^0.7.2"
yarl = "^1.9.4"
pytoml = "^0.1.21"
python-benedict = {extras = ["toml"], version = "^0.33.2"}

[tool.poetry.group.test.dependencies]
pytest = "^7.4.3"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
pytest-sugar = "^1.0.0"
pytest-clarity = "^1.0.1"
pytest-icdiff = "^0.9"
aioresponses = "^0.7.6"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.setuptools.packages.find]
where = ["."]  # list of folders that contain the packages (["."] by default)
include = ["ez_config_mgt*"]  # package names should match these glob patterns (["*"] by default)
exclude = ["test*"]  # exclude packages matching these glob patterns (empty by default)
namespaces = false  # to disable scanning PEP 420 namespaces (true by default)

[tool.bumpver]
current_version = "0.1.1"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "bump version {old_version} -> {new_version}"
commit          = true
tag             = true
push            = false

[tool.bumpver.file_patterns]
    "pyproject.toml" = [
        'current_version = "{version}"',
        'version = "{version}"',
    ]
    "ez_config_mgt/*.py" = ["{version}"]
    "tests/*.py" = ["{version}"]

[tool.isort]
profile                   = "black"
import_heading_stdlib     = "Standard library imports"
import_heading_thirdparty = "Third party imports"
import_heading_firstparty = "mongument imports"

[tool.mypy]
strict = true

    [[tool.mypy.overrides]]
    module                 = "ez_config_mgt"
    ignore_missing_imports = true

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

